* origin/master: (30 commits)
  Remove a test focus
  Fixed test cases
  Fix podspec not updated with latest version number
  Update for v4.5.0
  Follow style guide for boolean check
  Updated formatting of README.md
  Added steps for using the lock screen manager from SDL's bundle to the readme.
  Added check when trying to upload an appIcon for a head unit that does not support graphics.
  Update for v4.5.0
  Fixed issue with trying to upload a file that is nil will cause an infinite loop.
  Fixed an issue with initializer causing infinite loop.
  Improved readability for nil-checks.
  Fixed issue relating to SDLLocationDetails' searchAddress using the wrong key for object store.
  Fixed formatting for some interfaces.
  Fixed issue with LocationDetails' properties not properly becoming initialized.
  Fixed issue with waypoints not being correctly initialized as SDLLocationDetails
  Adding in nil checks for completion handlers of Streaming Media Manager.
  Updated description to return correct control data name.
  Update changelog for v4.5.0-beta.1
  Update for v4.5.0-beta.1 release
  ...

# Conflicts:
#	SmartDeviceLink/SDLDiagnosticMessage.m
#	SmartDeviceLink/SDLGetWaypointsResponse.m
#	SmartDeviceLink/SDLKeyboardProperties.h
#	SmartDeviceLink/SDLLocationDetails.m
#	SmartDeviceLink/SDLNames.h
#	SmartDeviceLink/SDLOnWaypointChange.m
#	SmartDeviceLink/SDLPerformInteraction.h
#	SmartDeviceLink/SDLPerformInteraction.m
#	SmartDeviceLink/SDLProxy.m
#	SmartDeviceLink/SDLPutFile.h
#	SmartDeviceLink/SDLReadDID.m
#	SmartDeviceLink/SDLSetGlobalProperties.h
#	SmartDeviceLink/SDLSetMediaClockTimer.h
#	SmartDeviceLink/SDLSoftButton.h
#	SmartDeviceLink/SDLUnsubscribeWaypoints.m
#	SmartDeviceLink/SDLUpdateTurnList.m
#	SmartDeviceLink/SDLV2ProtocolHeader.m
#	SmartDeviceLinkTests/RPCSpecs/RequestSpecs/SDLGetWaypointsSpec.m
#	SmartDeviceLinkTests/RPCSpecs/StructSpecs/SDLLocationDetailsSpec.m

+++ SmartDeviceLink/SDLAddSubMenu.h
@@ -18,9 +18,9 @@
  */
 @interface SDLAddSubMenu : SDLRPCRequest
 
-- (instancetype)initWithId:(UInt8)menuId menuName:(NSString *)menuName;
+- (instancetype)initWithId:(UInt32)menuId menuName:(NSString *)menuName;
 
-- (instancetype)initWithId:(UInt8)menuId menuName:(NSString *)menuName position:(UInt8)position;
+- (instancetype)initWithId:(UInt32)menuId menuName:(NSString *)menuName position:(UInt8)position;
 
 /**
  * @abstract a Menu ID that identifies a sub menu
